Update Sua ten bien, Nhieu hon 4 dung object/Post , them api thu hoi tai san

+++ RERPAPI.Model/Entities/TSAssetManagement.cs
@@ -72,7 +72,7 @@ public partial class TSAssetManagement
 
     public string? TSCodeNCC { get; set; }
 
-    public long? STT { get; set; }
+    public int? STT { get; set; }
 
 
     /// <summary>
